fix float double conversion
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@17864 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 5e8641a2b9b2c15778ee5474b5a78d5e96f8c58d

+++ Rec/Tf/PatAlgorithms/src/PatForwardTool.cpp
@@ -1,4 +1,4 @@
-// $Id: PatForwardTool.cpp,v 1.9 2008-05-14 17:22:18 mschille Exp $
+// $Id: PatForwardTool.cpp,v 1.10 2008-05-29 10:54:25 smenzeme Exp $
 // Include files
 
 // from Gaudi
@@ -495,7 +495,7 @@ void PatForwardTool::fillXList ( PatFwdTrackCandidate& track, double xMin, doubl
           PatFwdHit* hit = *itH;
           updateHitForTrack( hit, y0, ty);
 
-          if ( !hit->hit()->isYCompatible( (float)yRegion, (float)yCompat ) )
+          if ( !hit->hit()->isYCompatible( yRegion, yCompat ) )
             continue;
